\name{exp_calcAngleDist}
\alias{exp_calcAngleDist}
\title{Caclulate angle distance}
\author{Michael Lawrence <mflawren@fhcrc.org>}

\description{
Calculates angle distance between sample vectors \code{x} and \code{y}
}
\usage{exp_calcAngleDist(x, y)}
\arguments{
\item{x}{normally a column from the experimental data matrix}
\item{y}{normally a column from the experimental data matrix}
}

\details{}

\examples{}
\keyword{arith}
