\name{exp_calcCorrelationDist}
\alias{exp_calcCorrelationDist}
\title{Calculate correlation distance}
\author{Michael Lawrence <mflawren@fhcrc.org>}

\description{
Calculates correlation distance between \code{ent} and the other entities across
the conditions the samples (columns) in the data frame \code{ent_data}
}
\usage{exp_calcCorrelationDist(ent_data, ent)}
\arguments{
\item{ent_data}{a data frame of experimental data, according to exploRase conventions}
\item{ent}{the id of an entity that is compared to the entities in \code{ent_data}}
}

\details{}

\examples{}
\keyword{arith}
