\name{exp_calcDiff}
\alias{exp_calcDiff}
\title{Calculate difference}
\author{Michael Lawrence <mflawren@fhcrc.org>}

\description{
Just calculates \code{y} - \code{x}. You're probably better off just doing that.
}
\usage{exp_calcDiff(x, y)}
\arguments{
\item{x}{normally a column from the experimental data matrix}
\item{y}{normally a column from the experimental data matrix}
}

\details{}

\examples{}
\keyword{arith}
