\name{exp_calcEuclideanDist}
\alias{exp_calcEuclideanDist}
\title{Calculate Euclidean distance}
\author{Michael Lawrence <mflawren@fhcrc.org>}

\description{
Calculates Euclidean distance between \code{ent} and the other entities across 
the samples (columns) in the data frame \code{ent_data}
}
\usage{exp_calcEuclideanDist(ent_data, ent)}
\arguments{
\item{ent_data}{a data frame of experimental data, according to exploRase conventions}
\item{ent}{the id of an entity that is compared to the entities in \code{ent_data}}
}

\details{}

\examples{}
\keyword{arith}
