\name{exp_calcMahalanobisDist}
\alias{exp_calcMahalanobisDist}
\title{Calculate Mahalanobis distance}
\author{Michael Lawrence <mflawren@fhcrc.org>}

\description{
Calculates mahalanobis distance between the samples (columns) in the data frame \code{ent_data}
}
\usage{exp_calcMahalanobisDist(ent_data)}
\arguments{
\item{ent_data}{a data frame of experimental data, according to exploRase conventions}
}

\details{}

\examples{}
\keyword{arith}
