\name{exp_calcResiduals}
\alias{exp_calcResiduals}
\title{Calculate residuals}
\author{Michael Lawrence <mflawren@fhcrc.org>}

\description{
Calculates the residuals from a linear regression of sample \code{y} against sample \code{x}
}
\usage{exp_calcResiduals(x, y)}
\arguments{
\item{x}{normally a column from the experimental data matrix}
\item{y}{normally a column from the experimental data matrix}
}

\details{}

\examples{}
\keyword{regression}
