\name{exp_colorEntities}
\alias{exp_colorEntities}
\title{Coloring entities}
\author{Michael Lawrence <mflawren@fhcrc.org>}

\description{
Set the color for the specified entities of the current entity type in GGobi
}
\usage{exp_colorEntities(entities = getEntityIds(), color)}
\arguments{
\item{entities}{the entities to color}
\item{color}{the color index (as interpreted by GGobi) for the entities}
}

\details{}

\examples{}
\keyword{dynamic}
