\name{exp_designSelection}
\alias{exp_designSelection}
\title{Get selected conditions}
\author{Michael Lawrence <mflawren@fhcrc.org>}

\description{
Gets a vector of the names of the selected conditions in the condition list.
}
\usage{exp_designSelection()}
\arguments{

}
\value{A vector of condition names}
\details{}

\examples{}
\keyword{manip}
