\name{exp_entitiesInList}
\alias{exp_entitiesInList}
\title{Get the entities in a list}
\author{Michael Lawrence <mflawren@fhcrc.org>}

\description{
Gets the ID's of the entities in a specified list and of the specified types.
}
\usage{exp_entitiesInList(list, types = exp_entityTypes())}
\arguments{
\item{list}{The name of the entity list}
\item{types}{The types of entities to return}
}
\value{The entity ID's of the specified types belonging to the list}
\details{}

\examples{}
\keyword{manip}
