\name{exp_entityFrame}
\alias{exp_entityFrame}
\title{Get entity metadata}
\author{Michael Lawrence <mflawren@fhcrc.org>}

\description{
Retrieves the entity metadata table for the given type as a data frame
}
\usage{exp_entityFrame(ent_type = exp_entityType())}
\arguments{
\item{ent_type}{the entity type for which the metadata is retrieved}
}

\details{}

\examples{}
\keyword{manip}
