\name{exp_entitySelection}
\alias{exp_entitySelection}
\title{Get selected entities}
\author{Michael Lawrence <mflawren@fhcrc.org>}

\description{
Gets the ID's of the entities selected in the metadata table of the given type.
}
\usage{exp_entitySelection(ent_type = exp_entityType())}
\arguments{
\item{ent_type}{The entity type for which the selected is retrieved}
}
\value{The entity ID's}
\details{}

\examples{}
\keyword{manip}
