\name{exp_entityTypes}
\alias{exp_entityTypes}
\title{Get all entity types}
\author{Michael Lawrence <mflawren@fhcrc.org>}

\description{
Retrieves the identifiers for all entity types in exploRase.
}
\usage{exp_entityTypes()}
\arguments{

}
\value{Vector of entity type names}
\details{}

\examples{}
\keyword{manip}
