\name{exp_filterRules}
\alias{exp_filterRules}
\title{Get filter rules}
\author{Michael Lawrence <mflawren@fhcrc.org>}

\description{
Gets the filter rules for the metadata of the given type
}
\usage{exp_filterRules(ent_type = exp_entityType())}
\arguments{
\item{ent_type}{the entity type of the metadata being filtered by the rules}
}

\details{}

\examples{}
\keyword{manip}
