\name{exp_ggobi}
\alias{exp_ggobi}
\title{Get the GGobi session}
\author{Michael Lawrence <mflawren@fhcrc.org>}

\description{
Gets the GGobi session being used by exploRase. Will restart GGobi if it died
}
\usage{exp_ggobi()}
\arguments{

}
\value{The GGobi session}
\details{}

\examples{}
\keyword{dynamic}
