\name{exp_isRunning}
\alias{exp_isRunning}
\title{Check whether exploRase is running}
\author{Michael Lawrence <mflawren@fhcrc.org>}

\description{
Currently a hack that checks if the main window exists
}
\usage{exp_isRunning()}
\arguments{

}
\value{whether exploRase is currently running}
\details{}

\examples{}
\keyword{misc}
