\name{exp_listFrame}
\alias{exp_listFrame}
\title{Get entity lists}
\author{Michael Lawrence <mflawren@fhcrc.org>}

\description{
Gets the entity lists loaded in exploRase
}
\usage{exp_listFrame()}
\arguments{

}
\value{A data frame with a single column "name" holding the names of the lists}
\details{}

\examples{}
\keyword{manip}
