\name{exp_listSelection}
\alias{exp_listSelection}
\title{Get selected lists}
\author{Michael Lawrence <mflawren@fhcrc.org>}

\description{
Gets the names of the lists selected in the exploRase GUI.
}
\usage{exp_listSelection()}
\value{The names of the selected lists}
\details{}

\examples{}
\keyword{manip}
