\name{exp_loadDesign}
\alias{exp_loadDesign}
\title{Load experimental design}
\author{Michael Lawrence <mflawren@fhcrc.org>}

\description{
Loads a matrix describing the experimental conditions
}
\usage{exp_loadDesign(design_info, ent_type = exp_entityType())}
\arguments{
\item{design_info}{a data frame with conditions for rows and factors for cols}
\item{ent_type}{the type of entity to which this design applies. This links 
This links the design to the experimental data and entity information.}
}

\details{}

\examples{}
\keyword{manip}
