\name{exp_loadProject}
\alias{exp_loadProject}
\title{Load a project}
\author{Michael Lawrence <mflawren@fhcrc.org>}

\description{
Loads a project (a file system directory) into exploRase
}
\usage{exp_loadProject(project)}
\arguments{
\item{project}{The path to the directory holding the project}
}

\details{Loads all of the files in a specified directory, using their file extensions
to determine their purpose.}

\examples{}
\keyword{IO}
