\name{exp_newList}
\alias{exp_newList}
\title{Create an entity list}
\author{Michael Lawrence <mflawren@fhcrc.org>}

\description{
Forms the actual entity list matrix from the name, ents, and types
}
\usage{exp_newList(name, ents, types)}
\arguments{
\item{name}{The name of the entity list}
\item{ents}{The ID's of the entities in the list}
\item{types}{The types of the entities in the list}
}
\value{A matrix conforming to the structure for entity lists}
\details{}

\examples{}
\keyword{manip}
