\name{exp_removeFilterRules}
\alias{exp_removeFilterRules}
\title{Remove filter rules}
\author{Michael Lawrence <mflawren@fhcrc.org>}

\description{
Removes the filter rules with the given identifiers for the given type
}
\usage{exp_removeFilterRules(rules, ent_type = exp_entityType())}
\arguments{
\item{rules}{the identifiers of the rules to remove}
\item{ent_type}{the entity type of the metadata being filtered}
}

\details{}

\examples{}
\keyword{manip}
