\name{exp_showClustering}
\alias{exp_showClustering}
\title{Show a hierarchical clustering}
\author{Michael Lawrence <mflawren@fhcrc.org>}

\description{
Shows a hierarchical clustering using the "GGobi" dendrogram viewer.
}
\usage{exp_showClustering(ids, data, clustering)}
\arguments{
\item{ids}{The ids of the entities that were clustered}
\item{data}{The experimental data that was clustered}
\item{clustering}{the clustering, as returned by hclust()}
}

\details{}

\examples{}
\keyword{dynamic}
