\name{exp_toggleFilterRules}
\alias{exp_toggleFilterRules}
\title{Toggle filter rules}
\author{Michael Lawrence <mflawren@fhcrc.org>}

\description{
Toggles (activates or deactivates) the specified rules for the specified type.
}
\usage{exp_toggleFilterRules(rules, active = T, ent_type = exp_entityType())}
\arguments{
\item{rules}{The identifiers of the rules to toggle}
\item{active}{Whether to activate or deactivate the rules}
\item{ent_type}{The entity type of the metadata being filtered by the rules}
}

\details{}

\examples{}
\keyword{manip}
