\name{exp_entityType}
\alias{exp_entityType}
\title{Get the current entity type}
\author{Michael Lawrence <mflawren@fhcrc.org>}

\description{
Retrieves the entity type with its tab selected in the metadata notebook.
}
\usage{exp_entityType()}
\arguments{

}
\value{Name of the current entity type}
\details{}

\examples{}
\keyword{manip}
